Chore(Modularize the app): Transfer some elements from the main rebound.py file into modules
    - Create seperate modules for parts of the rebound.py file.
    - Make edits to the test file to accommodate new file structure
